+++ release_notes.md
@@ -1,3 +1,7 @@
+### 2.0.0
+* renamed 'append' to 'includeUtils'
+* added extend.getFirst
+
 ### 1.2.2
 * Appending new functions is now done explicitly to reduce chaos